steve v0.1.0 Steve.Storage behaviour

Defines a behaviour for storage adapters to implement.

Callbacks

ack!(arg1)

ack!(Steve.Job.t()) :: :ok | no_return()

Called after a job has been performed successfully.

dequeue(arg1, pos_integer)

dequeue(String.t(), pos_integer()) :: list()

Called to dequeue a specific amount of jobs.

enqueue(arg1, arg2)

enqueue(Steve.Job.t(), DateTime.t()) :: :ok | term()

Called to enqueue a given job to be performed at a specific time.

recover!(arg1)

recover!(String.t()) :: :ok | no_return()

Called upon start to hopefully recover orphaned jobs.

retry!(arg1)

retry!(Steve.Job.t()) :: :ok | no_return()

Called after an error is encountered when performing a job.