steve v0.1.0 Steve.Storage.Ecto

A storage adapter that uses Ecto to interact with a database.

Configuration

The application must be configured as shown below.

config :steve,
[
    storage: Steve.Storage.Ecto,
    repository: Example.Repo
]

Where Example.Repo should be a valid Ecto.Repo instance.

Migrations

One must also define the following migration within their application.

defmodule Example.Repo.Migrations.Steve do
    use Ecto.Migration

    defdelegate change, to: Steve.Storage.Ecto.Migration
end

Functions

ack!(job)

Called after a job has been performed successfully.

Callback implementation for Steve.Storage.ack!/1.

dequeue(queue, count)

Called to dequeue a specific amount of jobs.

Callback implementation for Steve.Storage.dequeue/2.

enqueue(job, time)

Called to enqueue a given job to be performed at a specific time.

Callback implementation for Steve.Storage.enqueue/2.

recover!(queue)

Called upon start to hopefully recover orphaned jobs.

Callback implementation for Steve.Storage.recover!/1.

retry!(job)

Called after an error is encountered when performing a job.

Callback implementation for Steve.Storage.retry!/1.