steve v0.1.0 Steve.Worker behaviour

Defines a behaviour for workers to implement.

Be sure to use this module for defaults to be injected.

Callbacks

backoff(retry)
(optional)

backoff(retry :: pos_integer()) :: non_neg_integer()

Called to determine a retry time offset in seconds.

Check the source code for the default implementation.

perform(...)

perform(... :: any()) :: any()

Called when a job is to be preformed.