steve v0.1.0 Steve.Worker behaviour

Defines a behaviour all workers should follow.

Be sure to use this module so that the defaults are injected.

Callbacks

backoff(retry) (optional)

backoff(retry :: pos_integer()) :: non_neg_integer()

Called to determine a retry time offset in seconds.

Check the source code for the default implementation.

perform(...)

perform(... :: any()) :: any()

Called when a job is to be performed.